+++ src/cli/serve.ts
@@ -10,6 +10,7 @@ import type { HanabiConfig } from "@/cli/config";
 import { bundleCss } from "@/cli/css";
 import { bundleWoff2 } from "@/cli/font";
 import { bundleHtml } from "@/cli/html";
+import { stringifyToHtml } from "@/cli/html";
 import { createAssetRouter, createPageRouter, createStaticRouter, withoutExt } from "@/cli/route";
 import type { Router } from "@/cli/route";
 import { bundleScriptEsbuild } from "@/cli/script";
@@ -17,7 +18,6 @@ import { Link, Script } from "@/lib/core/elements";
 import { clearStore, generateStore } from "@/lib/core/store";
 import type { Store } from "@/lib/core/store";
 import { contentType } from "@/lib/core/util";
-import { stringifyToHtml } from "@/lib/server/stringify";
 import { ErrorPage } from "@/page/error";
 import { hanabi_error_css } from "@/page/hanabi-error";
 import chokidar from "chokidar";
